requirements.txt
openai
anthropic
mirascope
python-dotenv
pydantic